+++ spinta/backends/postgresql/helpers/name.py
@@ -1,11 +1,30 @@
 from typing import Any
 
 from multipledispatch import dispatch
+from sqlalchemy.cimmutabledict import immutabledict
 
 from spinta.backends.constants import TableType
+from spinta.utils.sqlalchemy import Convention
 from spinta.backends.postgresql.helpers import get_pg_name
 from spinta.cli.helpers.migrate import MigrateRename
 from spinta.utils.itertools import ensure_list
+
+
+class _PgNamingConvention(str):
+
+    def __mod__(self, other) -> str:
+        name = super().__mod__(other)
+        return get_pg_name(name)
+
+
+# https://docs.sqlalchemy.org/en/14/core/constraints.html#configuring-a-naming-convention-for-a-metadata-collection
+PG_NAMING_CONVENTION = immutabledict({
+    Convention.IX: _PgNamingConvention("ix_%(table_name)s_%(column_0_N_name)s"),
+    Convention.UQ: _PgNamingConvention("uq_%(table_name)s_%(column_0_N_name)s"),
+    Convention.FK: _PgNamingConvention("fk_%(table_name)s_%(column_0_N_name)s"),
+    Convention.CK: _PgNamingConvention("ck_%(table_name)s_%(constraint_name)s"),
+    Convention.PK: _PgNamingConvention("pk_%(table_name)s"),
+})
 
 
 @dispatch(str, str)
@@ -35,6 +54,9 @@ def get_pg_changelog_name(table_name: str) -> str:
 
 
 def get_pg_file_name(table_name: str, arg: str = "") -> str:
+    if arg and not arg.startswith('/'):
+        arg = f'/{arg}'
+
     return get_pg_name(f"{table_name}{TableType.FILE.value}{arg}")
 
 
@@ -44,7 +66,10 @@ def get_pg_column_name(column_name: str) -> str:
 
 def get_pg_constraint_name(table_name: str, columns: Any) -> str:
     column_names = ensure_list(columns)
-    return get_pg_name(f"{table_name}_{'_'.join(column_names)}_key")
+    return PG_NAMING_CONVENTION[Convention.UQ] % {
+        "table_name": table_name,
+        "column_0_N_name": '_'.join(column_names)
+    }
 
 
 def get_pg_removed_name(name: str) -> str:
@@ -57,12 +82,18 @@ def get_pg_removed_name(name: str) -> str:
 
 
 def get_pg_foreign_key_name(table_name: str, column_name: str) -> str:
-    return get_pg_name(f"fk_{table_name}_{column_name.removeprefix('_')}")
+    return PG_NAMING_CONVENTION[Convention.FK] % {
+        "table_name": table_name,
+        "column_0_N_name": column_name.removeprefix('_')
+    }
 
 
 def get_pg_index_name(table_name: str, columns: Any) -> str:
     column_names = ensure_list(columns)
-    return get_pg_name(f"ix_{table_name}_{'_'.join(column_names)}")
+    return PG_NAMING_CONVENTION[Convention.IX] % {
+        "table_name": table_name,
+        "column_0_N_name": '_'.join(column_names)
+    }
 
 
 def is_removed(name: str) -> bool: